requirements.txt
# Core FastAPI framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# LangChain stack - using compatible versions
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.30

# Vector database and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# PDF processing
pypdf>=3.0.0

# Web search integrations (optional - mock is default)
tavily-python>=0.3.0
duckduckgo-search>=4.0.0

# Memory and caching
redis>=5.0.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
loguru>=0.7.0

# HTTP client
httpx>=0.25.0

# Data processing
numpy>=1.24.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Monitoring and metrics
prometheus-client>=0.17.0
prometheus-fastapi-instrumentator>=6.1.0